# docs/openapi.yaml
openapi: 3.0.3
info:
  title: mStream API
  version: "1"
  description: |
    REST API for the [mStream](https://github.com/IrosTheBeggar/mStream) music streaming server.

    ## Authentication

    Most endpoints require a JWT bearer token. Tokens are issued by `POST /api/v1/auth/login` and
    can be supplied via any of (checked in order):

    1. Request body field `token`
    2. Query parameter `token`
    3. HTTP header `x-access-token`
    4. Cookie `x-access-token`

    ### Public mode

    If no users exist in the database, all requests are authenticated as a virtual public user with
    full library access — no token required. This is the default for fresh installs.

    ### Special tokens

    - **Admin tokens** have `admin: true` in the payload and are required for every `/api/v1/admin/*` endpoint.
    - **Jukebox tokens** (`jukebox: true`) are short-lived and bound to an active remote-control session.
    - **Shared-playlist tokens** (`shareToken: true`) restrict file access to the tracks in one specific shared playlist.
    - **Federation invite tokens** (`invite: true`) can only call `POST /api/v1/federation/invite/exchange`.

    ## Errors

    Errors use the shape `{ "error": "<message>" }`. Status codes:

    | Code | Meaning |
    |------|---------|
    | 400  | Bad request (invalid input) |
    | 401  | Unauthenticated |
    | 403  | Forbidden (Joi validation error, missing permission) |
    | 404  | Resource not found |
    | 405  | Admin API locked |
...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass the JWT in the `x-access-token` header. Despite the OpenAPI "bearer" label,
        the value does NOT include a `Bearer ` prefix — it's the raw token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: x-access-token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: Authentication Error
...
  responses:
...
    Unauthorized:
      description: Auth required or token invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/db/waveform:
    get:
      tags: [Waveform]
      summary: Get waveform peaks for a track
      description: 800 bars, 0–255 each. Generated on demand and cached.
      parameters:
        - name: filepath
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Waveform data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  waveform:
                    type: array
                    items: { type: integer, minimum: 0, maximum: 255 }
                    minItems: 800
                    maxItems: 800
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ── Cue Points ───────────────────────────────────────────────────────────
...